+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
-## Version 0.1.1-SNAPSHOT
-* Make it possible to use the plugin with Java 7
+## Version 0.1.1 *(2015-11-29)*
+* Make it possible to use the plugin with Java 7 (fixes #11)
 * Provide default name patterns for classes to be excluded from report generation.
 Currently, these patterns include *Android*, *Dagger 2* and *Butter Knife* generated classes.
 * Allow user to toggle report generation for specified type